Change: move sql and sqlc directories. centralize sqlc config
sql queries for go could now be generated/re-generated with one command
from backend's root directory with "sqlc generate" command"

+++ backend/monolith/internal/domains/driver/infrastructure/postgres/sqlc/models.go
@@ -28,7 +28,7 @@ type Driver struct {
 }
 
 type DriverRealtime struct {
-	DriverID            int64
+	DriverID            uint32
 	UpdatedAt           *time.Time
 	RealtimeLocation    interface{}
 	AverageSpeed        *float32
